fix(channels): move session-recovery helper to @openhermit/shared
The stale-session recovery helper (added in #222) lived in @openhermit/sdk,
which channel packages depend on via a caret range (^0.6.0). The workspace
SDK is 0.7.0, which ^0.6.0 does NOT satisfy, so npm installs a nested
published 0.6.x copy under each channel. Those channels resolved
@openhermit/sdk to the nested 0.6.x (no openSessionWithFreshFallback
export) → the whole channel manifest failed to load → telegram/slack/
discord/wechat showed "no manifest registered for channel" and couldn't
start. (whatsapp escaped it — it pins ^0.7.0.) Rebuilding/​bumping didn't
help reliably: npm won't prune an already-installed nested copy, and the
amiko deploy reuses a persistent volume (git reset + npm install, no clean
install).

Move the helper to @openhermit/shared instead. Channels pin
@openhermit/shared to an EXACT workspace version, so it is always resolved
to the workspace copy (no nested copies exist) — the new export is reliably
visible regardless of npm hoisting/prune behavior. Verified: all six
channels load even while a stale nested @openhermit/sdk 0.6.1 is still
present.

- @openhermit/shared: add isSessionNotFoundError + openSessionWithFreshFallback
  (with unit tests); revert the same from @openhermit/sdk.
- All six bridges import the helper from @openhermit/shared.

Refs #210, #222.
