+++ apps/channels/wechat/src/bridge.ts
@@ -8,14 +8,14 @@
  */
 import { randomUUID } from 'node:crypto';
 
-import { AgentLocalClient, parseSseFrames, openSessionWithFreshFallback } from '@openhermit/sdk';
+import { AgentLocalClient, parseSseFrames } from '@openhermit/sdk';
 import type {
   ChannelMessageAction,
   ChannelOutbound,
   ChannelOutboundResult,
   OutboundSession,
 } from '@openhermit/protocol';
-import { stripSilenceTokens } from '@openhermit/shared';
+import { stripSilenceTokens, openSessionWithFreshFallback } from '@openhermit/shared';
 
 import { sendMessage } from './ilink/api.js';
 import {
