+++ apps/channels/whatsapp/src/bridge.ts
@@ -1,11 +1,11 @@
-import { AgentLocalClient, parseSseFrames, openSessionWithFreshFallback } from '@openhermit/sdk';
+import { AgentLocalClient, parseSseFrames } from '@openhermit/sdk';
 import type {
   ChannelMessageAction,
   ChannelOutbound,
   ChannelOutboundResult,
   OutboundSession,
 } from '@openhermit/protocol';
-import { stripSilenceTokens } from '@openhermit/shared';
+import { stripSilenceTokens, openSessionWithFreshFallback } from '@openhermit/shared';
 
 import { formatAgentResponse } from './formatting.js';
 import {
